+++ pdf/lib/src/pdf/obj/font.dart
@@ -194,7 +194,7 @@ See https://github.com/DavBfr/dart_pdf/wiki/Fonts-Management
   /// Draw some text
   void putText(PdfStream stream, String text) {
     try {
-      PdfString(latin1.encode(text), PdfStringFormat.litteral).output(stream);
+      PdfString(latin1.encode(text), PdfStringFormat.literal).output(stream);
     } catch (_) {
       assert(() {
         print(_cannotDecodeMessage);